Refactor Markdown table contents generation to generic Markup function
